+++ spec/dummy/app/models/wine.rb
@@ -14,6 +14,8 @@ class Wine < ActiveRecord::Base
 
   validates :name_en, :name_zh, :name_ja, presence: true
 
+  translate :name
+
   def fae_display_field
     name_en
   end